refactor: formatTemplates에 타입 추론 추가하여 date.ts를 src/utils/로 이동
