+++ src/components/CalendarDialog/DateField.tsx
@@ -1,7 +1,7 @@
 import { IcCalenderLine } from '@yourssu/design-system-react';
 import { ComponentProps, forwardRef } from 'react';
 
-import { formatTemplates } from '@/components/CalendarDialog/date';
+import { formatTemplates } from '@/utils/date';
 import { DateFieldContainer } from '@/components/CalendarDialog/DateField.style';
 
 interface DateFieldProps extends ComponentProps<'div'> {
